requirements.txt
aiogram==3.4.1
openai>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
